requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.5
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.5.0
